requirements.txt
django-allauth>=0.54.0
